IPI scores

Estimate IPI scores for PHAB using station and physical habitat data

IPI(stations, phab, qa = TRUE, allerr = TRUE, log = FALSE)

Arguments

stations
data.frame of input station data

phab
data.frame of input physical habitat data

qa
logical indicating of quality assurance columns are appended to the output

allerr
logical indicating if all errors are returned or the first encountered

log
logical indicating if errors are printed to log in the workign directory, applies only if allerr = TRUE

Value

A results data.frame of IPI and metric scores, quality assurance results are returned if qa = TRUE (default)

Examples

IPI(stations, phab)

#>   StationCode SampleDate       PHAB_SampleID  IPI IPI_percentile Ev_FlowHab
#> 1   105PS0107  9/14/2009 105PS0107_9/14/2009 1.16           0.90       0.85
#> 2   205PS0157  6/19/2012 205PS0157_6/19/2012 1.04           0.62       0.96
#> 3   305PS0057  6/16/2009 305PS0057_6/16/2009 0.79           0.04       0.50
#> 4   504PS0147  6/23/2008 504PS0147_6/23/2008 0.78           0.03       0.28
#> 5   632PS0007  7/23/2008 632PS0007_7/23/2008 1.10           0.79       0.90
#> 6   901PS0057  5/14/2008 901PS0057_5/14/2008 1.08           0.74       0.70
#>   Ev_FlowHab_score H_AqHab H_AqHab_pred H_AqHab_score H_SubNat H_SubNat_score
#> 1             0.89    1.59         1.11          1.00     1.57           0.83
#> 2             1.00    1.42         1.35          0.80     1.41           0.74
#> 3             0.51    1.32         1.42          0.70     0.49           0.26
#> 4             0.28    1.24         1.30          0.72     0.98           0.52
#> 5             0.95    1.51         1.41          0.82     1.80           0.95
#> 6             0.73    1.52         1.38          0.84     1.80           0.95
#>   PCT_SAFN PCT_RC PCT_SAFN_pred PCT_SAFN_score XCMG XCMG_pred XCMG_score IPI_qa
#> 1        6      0         24.60           1.00   99     93.64       0.78   1.00
#> 2       51      2         22.06           0.40  131    104.72       0.90   1.00
#> 3       83      0         29.51           0.12  152    106.05       1.00   1.00
#> 4        1      0         34.38           1.00   55     95.41       0.51   1.00
#> 5       14      0         13.49           0.79  126    123.10       0.76   1.00
#> 6       40      3         34.46           0.69  122    102.16       0.86   0.95
#>   Ev_FlowHab_qa H_AqHab_qa H_SubNat_qa PCT_SAFN_qa XCMG_qa
#> 1             1          1           1           1    1.00
#> 2             1          1           1           1    1.00
#> 3             1          1           1           1    1.00
#> 4             1          1           1           1    1.00
#> 5             1          1           1           1    1.00
#> 6             1          1           1           1    0.95